+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 BUG FIXES:
 * clickhouse: fixed resetting kafka_topics timeout settings to zero when adding a new topic
 * storage: fix lifecycle state upgrader
+* mongodb: fixed mongocfg/mongos terraform configs behaviour with mongoinfra hosts
 
 ## 0.108.0 (February 14, 2024)
 FEATURES: